chore(mmex): big file split
