+++ src/mmcheckingpanel.h
@@ -33,180 +33,6 @@ class mmFilterTransactionsDialog;
 class mmGUIFrame;
 //----------------------------------------------------------------------------
 
-class TransactionListCtrl : public mmListCtrl
-{
-public:
-
-    TransactionListCtrl(mmCheckingPanel* cp
-        , wxWindow* parent
-        , const wxWindowID id = wxID_ANY);
-
-    ~TransactionListCtrl();
-
-    void createColumns(mmListCtrl &lst);
-    enum EIcons //m_imageList
-    {
-        ICON_RECONCILED,
-        ICON_VOID,
-        ICON_FOLLOWUP,
-        ICON_NONE,
-        ICON_DUPLICATE,
-        ICON_DESC,
-        ICON_ASC,
-    };
-
-    enum EColumn
-    {
-        COL_IMGSTATUS = 0,
-        COL_ID,
-        COL_DATE,
-        COL_NUMBER,
-        COL_PAYEE_STR,
-        COL_STATUS,
-        COL_CATEGORY,
-        COL_WITHDRAWAL,
-        COL_DEPOSIT,
-        COL_BALANCE,
-        COL_NOTES,
-        COL_MAX, // number of columns
-        COL_DEF_SORT = COL_DATE
-    };
-    EColumn toEColumn(long col)
-    {
-        EColumn res = COL_DEF_SORT;
-        if (col >= 0 && col < COL_MAX) res = static_cast<EColumn>(col);
-
-        return res;
-    }
-
-    EColumn g_sortcol; // index of column to sort
-    EColumn m_prevSortCol;
-    bool g_asc; // asc\desc sorting
-
-    bool getSortOrder() const { return m_asc; }
-    EColumn getSortColumn() const { return m_sortCol; }
-
-    void setSortOrder(bool asc) { m_asc = asc; }
-    void setSortColumn(EColumn col) { m_sortCol = col; }
-
-    void setColumnImage(EColumn col, int image);
-
-    void OnNewTransaction(wxCommandEvent& event);
-    void OnNewTransferTransaction(wxCommandEvent& event);
-    void OnDeleteTransaction(wxCommandEvent& event);
-    void OnEditTransaction(wxCommandEvent& event);
-    void OnDuplicateTransaction(wxCommandEvent& event);
-    void OnSetUserColour(wxCommandEvent& event);
-    void OnMoveTransaction(wxCommandEvent& event);
-    void OnOpenAttachment(wxCommandEvent& event);
-    /// Displays the split categories for the selected transaction
-    void OnViewSplitTransaction(wxCommandEvent& event);
-    void OnOrganizeAttachments(wxCommandEvent& event);
-    void OnCreateReoccurance(wxCommandEvent& event);
-    void refreshVisualList(int trans_id = -1, bool filter = true);
-    long m_selectedIndex;
-    long m_selectedForCopy; //The transaction ID if selected for copy
-    long m_selectedID; //Selected transaction ID
-
-protected:
-    /* Sort Columns */
-    virtual void OnColClick(wxListEvent& event);
-
-private:
-    enum
-    {
-        MENU_TREEPOPUP_MARKRECONCILED = wxID_HIGHEST + 150,
-        MENU_TREEPOPUP_MARKUNRECONCILED,
-        MENU_TREEPOPUP_MARKVOID,
-        MENU_TREEPOPUP_MARK_ADD_FLAG_FOLLOWUP,
-        MENU_TREEPOPUP_MARKDUPLICATE,
-        MENU_TREEPOPUP_MARKDELETE,
-        MENU_TREEPOPUP_MARKRECONCILED_ALL,
-        MENU_TREEPOPUP_MARKUNRECONCILED_ALL,
-        MENU_TREEPOPUP_MARKVOID_ALL,
-        MENU_TREEPOPUP_MARK_ADD_FLAG_FOLLOWUP_ALL,
-        MENU_TREEPOPUP_MARKDUPLICATE_ALL,
-        MENU_TREEPOPUP_DELETE_VIEWED,
-        MENU_TREEPOPUP_DELETE_FLAGGED,
-        MENU_TREEPOPUP_DELETE_UNRECONCILED,
-
-        MENU_TREEPOPUP_VIEW_SPLIT_CATEGORIES,
-        MENU_TREEPOPUP_ORGANIZE_ATTACHMENTS,
-        MENU_TREEPOPUP_CREATE_REOCCURANCE,
-        MENU_SUBMENU_MARK_ALL,
-
-        MENU_VIEW_,
-        MENU_VIEW_DELETE_TRANS,
-        MENU_VIEW_DELETE_FLAGGED,
-
-        MENU_ON_COPY_TRANSACTION,
-        MENU_ON_PASTE_TRANSACTION,
-        MENU_ON_NEW_TRANSACTION,
-        MENU_ON_DUPLICATE_TRANSACTION,
-
-        MENU_ON_SET_UDC0, //Default colour
-        MENU_ON_SET_UDC1, //User defined colour 1
-        MENU_ON_SET_UDC2, //User defined colour 2
-        MENU_ON_SET_UDC3, //User defined colour 3
-        MENU_ON_SET_UDC4, //User defined colour 4
-        MENU_ON_SET_UDC5, //User defined colour 5
-        MENU_ON_SET_UDC6, //User defined colour 6
-        MENU_ON_SET_UDC7, //User defined colour 7
-
-        MENU_TREEPOPUP_NEW_WITHDRAWAL,
-        MENU_TREEPOPUP_NEW_DEPOSIT,
-        MENU_TREEPOPUP_NEW_TRANSFER,
-        MENU_TREEPOPUP_EDIT2,
-        MENU_TREEPOPUP_MOVE2,
-        MENU_TREEPOPUP_DELETE2,
-        ID_PANEL_CHECKING_STATIC_BITMAP_VIEW,
-    };
-private:
-    DECLARE_NO_COPY_CLASS(TransactionListCtrl)
-    wxDECLARE_EVENT_TABLE();
-
-    mmCheckingPanel* m_cp;
-
-    wxSharedPtr<wxListItemAttr> m_attr1;  // style1
-    wxSharedPtr<wxListItemAttr> m_attr2;  // style2
-    wxSharedPtr<wxListItemAttr> m_attr3;  // style, for future dates
-    wxSharedPtr<wxListItemAttr> m_attr4;  // style, for future dates
-    wxSharedPtr<wxListItemAttr> m_attr11; // user defined style 1
-    wxSharedPtr<wxListItemAttr> m_attr12; // user defined style 2
-    wxSharedPtr<wxListItemAttr> m_attr13; // user defined style 3
-    wxSharedPtr<wxListItemAttr> m_attr14; // user defined style 4
-    wxSharedPtr<wxListItemAttr> m_attr15; // user defined style 5
-    wxSharedPtr<wxListItemAttr> m_attr16; // user defined style 6
-    wxSharedPtr<wxListItemAttr> m_attr17; // user defined style 7
-
-    /* required overrides for virtual style list control */
-    virtual wxString OnGetItemText(long item, long column) const;
-    virtual int OnGetItemColumnImage(long item, long column) const;
-    virtual wxListItemAttr* OnGetItemAttr(long item) const;
-
-    void OnMouseRightClick(wxMouseEvent& event);
-    void OnListLeftClick(wxMouseEvent& event);
-    void OnListItemSelected(wxListEvent& event);
-    void OnListItemActivated(wxListEvent& event);
-    void OnMarkTransaction(wxCommandEvent& event);
-    void OnMarkAllTransactions(wxCommandEvent& event);
-    void OnListKeyDown(wxListEvent& event);
-    void OnChar(wxKeyEvent& event);
-    void OnCopy(wxCommandEvent& WXUNUSED(event));
-    void OnPaste(wxCommandEvent& WXUNUSED(event));
-    int OnPaste(Model_Checking::Data* tran);
-
-    bool TransactionLocked(const wxString& transdate);
-private:
-    /* The topmost visible item - this will be used to set
-    where to display the list again after refresh */
-    long m_topItemIndex;
-    EColumn m_sortCol;
-    wxString m_today;
-};
-
-//----------------------------------------------------------------------------
-
 class mmCheckingPanel : public mmPanelBase
 {
 public:
@@ -229,6 +55,8 @@ public:
     void RefreshList(int transID = -1);
 
     wxString BuildPage() const;
+    /* Getter for Virtual List Control */
+    const wxString getItem(long item, long column);
 
 private:
     enum
@@ -310,7 +138,6 @@ private:
     Model_Account::Data* m_account;
     Model_Currency::Data* m_currency;
     wxScopedPtr<wxImageList> m_imageList;
-    Model_Checking::Full_Data_Set m_trans;
 
     void initViewTransactionsHeader();
     void initFilterSettings();
@@ -318,7 +145,6 @@ private:
     void sortTable();
     void filterTable();
     void updateTable();
-    void markSelectedTransaction(int trans_id);
     void CreateControls();
 
     bool Create(
@@ -343,16 +169,10 @@ private:
     void OnViewPopupSelected(wxCommandEvent& event);
     void OnSearchTxtEntered(wxCommandEvent& event);
 
-    void DeleteViewedTransactions();
-    void DeleteFlaggedTransactions(const wxString& status);
-
     /* updates the checking panel data */
     void showTips();
     void updateExtraTransactionData(int selIndex);
     wxString GetPanelTitle(const Model_Account::Data& account) const;
-
-    /* Getter for Virtual List Control */
-    const wxString getItem(long item, long column);
 
     static void mmPlayTransactionSound();
     mmGUIFrame* m_frame;
