+++ src/model/Model_Account.h
@@ -112,6 +112,12 @@ public:
 
     static wxDateTime DateOf(const wxString& date_str);
     static bool BoolOf(int value);
+    static bool is_positive(int value);
+    static wxDateTime get_date_by_string(const wxString& date_str);
+
 };
+
+inline wxDateTime Model_Account::get_date_by_string(const wxString& date_str) { return Model::to_date(date_str); }
+inline bool Model_Account::is_positive(int value) { return value > 0 ? true : false; }
 
 #endif //
